[PR] main <- develop

+++ JeonsiLog/app/src/main/java/com/example/jeonsilog/view/mypage/MyPageRvAdapter.kt
@@ -1,16 +1,23 @@
 package com.example.jeonsilog.view.mypage
 
+import android.content.Context
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.content.res.AppCompatResources
 import androidx.recyclerview.widget.RecyclerView
-import com.example.jeonsilog.data.remote.dto.interest.GetInterestInformationEntity
-import com.example.jeonsilog.data.remote.dto.rating.GetMyRatingsDataEntity
-import com.example.jeonsilog.data.remote.dto.review.GetReviewsDataEntity
+import com.bumptech.glide.load.resource.bitmap.CenterCrop
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.example.jeonsilog.R
+import com.example.jeonsilog.data.remote.dto.interest.GetInterestEntity
+import com.example.jeonsilog.data.remote.dto.rating.GetMyRatingsEntity
+import com.example.jeonsilog.data.remote.dto.review.GetReviewsEntity
 import com.example.jeonsilog.databinding.ItemMyPageInterestBinding
 import com.example.jeonsilog.databinding.ItemMyPageRatingBinding
 import com.example.jeonsilog.databinding.ItemMyPageReviewBinding
 import com.example.jeonsilog.repository.interest.InterestRepositoryImpl
+import com.example.jeonsilog.view.MainActivity
+import com.example.jeonsilog.view.exhibition.ExtraActivity
 import com.example.jeonsilog.widget.utils.GlideApp
 import com.example.jeonsilog.widget.utils.GlobalApplication.Companion.encryptedPrefs
 import kotlinx.coroutines.CoroutineScope
@@ -18,23 +25,25 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import java.lang.IllegalArgumentException
 
-class MyPageRvAdapter<T>(private val list: MutableList<T>, private val type: Int): RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+class MyPageRvAdapter<T>(private val list: MutableList<T>, private val type: Int, private val context: Context): RecyclerView.Adapter<RecyclerView.ViewHolder>() {
 
     inner class TypeRatingViewHolder(private val binding: ItemMyPageRatingBinding): RecyclerView.ViewHolder(binding.root){
-        fun bind(data: GetMyRatingsDataEntity){
+        fun bind(data: GetMyRatingsEntity){
             binding.tvMypageRatingItemTitle.text = data.exhibitionName
             binding.rbMypageRatingItemRating.rating = data.rate.toFloat()
 
-            // 클릭리스너 - 해당 전시회 상세 페이지로 이동
             itemView.setOnClickListener {
-
+                if (context.javaClass.simpleName == "MainActivity"){
+                    (context as MainActivity).loadExtraActivity(type = 0, newTargetId = data.exhibitionId)
+                }else if(context.javaClass.simpleName=="ExtraActivity"){
+                    (context as ExtraActivity).loadExtraActivity(type = 0, newTargetId = data.exhibitionId)
+                }
             }
-
         }
     }
 
     inner class TypeReviewViewHolder(private val binding: ItemMyPageReviewBinding): RecyclerView.ViewHolder(binding.root){
-        fun bind(data: GetReviewsDataEntity) {
+        fun bind(data: GetReviewsEntity) {
             GlideApp.with(binding.ivMypageReviewExhibitionImg)
                 .load(data.exhibitionImgUrl)
                 .centerCrop()
@@ -43,50 +52,66 @@ class MyPageRvAdapter<T>(private val list: MutableList<T>, private val type: Int
             binding.tvMypageReviewTitle.text = data.exhibitionName
             binding.tvMypageReviewContent.text = data.contents
 
-            // 클릭리스너 - 해당 전시회 상세 페이지
             itemView.setOnClickListener {
-
+                if (context.javaClass.simpleName == "MainActivity"){
+                    (context as MainActivity).loadExtraActivity(type = 0, newTargetId = data.exhibitionId)
+                }else if(context.javaClass.simpleName=="ExtraActivity"){
+                    (context as ExtraActivity).loadExtraActivity(type = 0, newTargetId = data.exhibitionId)
+                }
             }
         }
     }
 
     inner class TypeInterestViewHolder(private val binding: ItemMyPageInterestBinding): RecyclerView.ViewHolder(binding.root){
-        fun bind(data: GetInterestInformationEntity) {
+        fun bind(data: GetInterestEntity) {
             GlideApp.with(binding.ivMypageInterestExhibitionImg)
                 .load(data.imageUrl)
-                .centerCrop()
+                .transform(CenterCrop(), RoundedCorners(16))
                 .into(binding.ivMypageInterestExhibitionImg)
 
             binding.tvMypageInterestTitle.text = data.exhibitionName
-            binding.tvMypageInterestAddress.text = data.placeName
 
-            binding.tvMypageInterestKeywordBefore.visibility = View.GONE
-            binding.tvMypageInterestKeywordOn.visibility = View.GONE
-            binding.tvMypageInterestKeywordFree.visibility = View.GONE
+            binding.tvMypageInterestAddress.text = placeName(data.placeName)
 
-            if(data.operatingKeyword == "시작전"){
-                binding.tvMypageInterestKeywordBefore.visibility = View.VISIBLE
-            }
-            if(data.operatingKeyword == "전시중"){
-                binding.tvMypageInterestKeywordOn.visibility = View.VISIBLE
-            }
-            if(data.priceKeyword == "무료"){
-                binding.tvMypageInterestKeywordFree.visibility = View.VISIBLE
-            }
-
-            binding.ibMypageInterest.setOnClickListener {
-                list.removeAt(adapterPosition)
-                notifyItemRemoved(adapterPosition)
-
-                // 서버에 즐겨찾기 해제 요청
-                CoroutineScope(Dispatchers.IO).launch{
-                    InterestRepositoryImpl().deleteInterest(encryptedPrefs.getAT(), data.exhibitionId)
+            when(data.operatingKeyword){
+                "ON_DISPLAY" -> binding.tvMypageInterestKeywordOn.visibility = View.VISIBLE
+                "BEFORE_DISPLAY" -> binding.tvMypageInterestKeywordBefore.visibility = View.VISIBLE
+                else -> {
+                    binding.tvMypageInterestKeywordBefore.visibility = View.GONE
+                    binding.tvMypageInterestKeywordOn.visibility = View.GONE
                 }
             }
 
-            // 클릭리스너 - 해당 전시회 상세 페이지
-            itemView.setOnClickListener {
+            when(data.priceKeyword){
+                "FREE" -> binding.tvMypageInterestKeywordFree.visibility = View.VISIBLE
+                else -> binding.tvMypageInterestKeywordFree.visibility = View.GONE
+            }
 
+            var state = true
+            binding.ibMypageInterest.background = AppCompatResources.getDrawable(itemView.context, R.drawable.ic_interest_active)
+
+            binding.ibMypageInterest.setOnClickListener {
+                if(state){
+                    binding.ibMypageInterest.background = AppCompatResources.getDrawable(itemView.context, R.drawable.ic_interest_inactive)
+                    CoroutineScope(Dispatchers.IO).launch{
+                        InterestRepositoryImpl().deleteInterest(encryptedPrefs.getAT(), data.exhibitionId)
+                    }
+                } else {
+                    binding.ibMypageInterest.background = AppCompatResources.getDrawable(itemView.context, R.drawable.ic_interest_active)
+                    CoroutineScope(Dispatchers.IO).launch{
+                        InterestRepositoryImpl().postInterest(encryptedPrefs.getAT(), data.exhibitionId)
+                    }
+                }
+
+                state = !state
+            }
+
+            itemView.setOnClickListener {
+                if (context.javaClass.simpleName == "MainActivity"){
+                    (context as MainActivity).loadExtraActivity(type = 0, newTargetId = data.exhibitionId)
+                }else if(context.javaClass.simpleName=="ExtraActivity"){
+                    (context as ExtraActivity).loadExtraActivity(type = 0, newTargetId = data.exhibitionId)
+                }
             }
         }
     }
@@ -129,17 +154,17 @@ class MyPageRvAdapter<T>(private val list: MutableList<T>, private val type: Int
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         when (this.type) {
             0 -> {
-                val ratingData = list[position] as GetMyRatingsDataEntity
+                val ratingData = list[position] as GetMyRatingsEntity
                 holder as MyPageRvAdapter<*>.TypeRatingViewHolder
                 holder.bind(ratingData)
             }
             1 -> {
-                val reviewData = list[position] as GetReviewsDataEntity
+                val reviewData = list[position] as GetReviewsEntity
                 holder as MyPageRvAdapter<*>.TypeReviewViewHolder
                 holder.bind(reviewData)
             }
             2 -> {
-                val interestData = list[position] as GetInterestInformationEntity
+                val interestData = list[position] as GetInterestEntity
                 holder as MyPageRvAdapter<*>.TypeInterestViewHolder
                 holder.bind(interestData)
             }
@@ -149,5 +174,18 @@ class MyPageRvAdapter<T>(private val list: MutableList<T>, private val type: Int
 
     override fun getItemViewType(position: Int): Int {
         return type
+    }
+
+    private fun placeName(input: String?): String {
+        return if(input.isNullOrEmpty()){
+            ""
+        } else {
+            val index = input.indexOf('(')
+            if (index != -1) {
+                input.substring(0, index)
+            } else {
+                input
+            }
+        }
     }
 }
